refactor(library): split PlaylistSelection into LibraryPage and DrawerLibrary (#930)
Extract shared state/context logic into useLibraryRoot hook and create
two focused composition roots: LibraryPage (standalone full-screen) and
DrawerLibrary (drawer-embedded). Remove $inDrawer conditionals from
Container/SelectionCard styled-components in styled.layout.ts.
