+++ src/components/__tests__/PlaylistSelection.test.tsx
@@ -3,7 +3,7 @@ import { render, screen, fireEvent } from '@testing-library/react';
 import { vi, describe, it, expect, beforeEach } from 'vitest';
 import { ThemeProvider } from 'styled-components';
 import { theme } from '@/styles/theme';
-import PlaylistSelection from '../PlaylistSelection';
+import { LibraryPage, DrawerLibrary } from '../PlaylistSelection';
 import { TestWrapper } from '@/test/testWrappers';
 import { makePlaylistInfo, makeAlbumInfo } from '@/test/fixtures';
 import { LIKED_SONGS_ID } from '@/constants/playlist';
@@ -97,12 +97,24 @@ function setMockLibrarySync(overrides?: Record<string, unknown>) {
   } as ReturnType<typeof useLibrarySync>);
 }
 
-function renderPlaylistSelection(props?: Partial<Parameters<typeof PlaylistSelection>[0]>) {
+function renderLibraryPage(props?: Partial<Parameters<typeof LibraryPage>[0]>) {
   const onPlaylistSelect = vi.fn();
   const result = render(
     <ThemeProvider theme={theme}>
       <TestWrapper>
-        <PlaylistSelection onPlaylistSelect={onPlaylistSelect} {...props} />
+        <LibraryPage onPlaylistSelect={onPlaylistSelect} {...props} />
+      </TestWrapper>
+    </ThemeProvider>
+  );
+  return { ...result, onPlaylistSelect };
+}
+
+function renderDrawerLibrary(props?: Partial<Parameters<typeof DrawerLibrary>[0]>) {
+  const onPlaylistSelect = vi.fn();
+  const result = render(
+    <ThemeProvider theme={theme}>
+      <TestWrapper>
+        <DrawerLibrary onPlaylistSelect={onPlaylistSelect} {...props} />
       </TestWrapper>
     </ThemeProvider>
   );
@@ -115,14 +127,14 @@ describe('PlaylistSelection', () => {
   });
 
   it('renders Playlists and Albums tab buttons', () => {
-    renderPlaylistSelection();
+    renderLibraryPage();
     expect(screen.getByRole('button', { name: /playlists/i })).toBeTruthy();
     expect(screen.getByRole('button', { name: /albums/i })).toBeTruthy();
   });
 
   it('switching to Albums tab shows the album grid', () => {
     // #given
-    renderPlaylistSelection();
+    renderLibraryPage();
 
     // #when
     fireEvent.click(screen.getByRole('button', { name: /albums/i }));
@@ -134,7 +146,7 @@ describe('PlaylistSelection', () => {
 
   it('search input filters playlist list by name (case-insensitive)', () => {
     // #given
-    renderPlaylistSelection();
+    renderLibraryPage();
     const searchInput = screen.getByPlaceholderText('Search playlists...');
 
     // #when
@@ -146,14 +158,14 @@ describe('PlaylistSelection', () => {
   });
 
   it('pinned items appear before unpinned items', () => {
-    renderPlaylistSelection();
+    renderLibraryPage();
     const items = screen.getAllByText(/tracks/);
     expect(items.length).toBeGreaterThan(0);
   });
 
   it('clicking a playlist calls onPlaylistSelect with the correct playlist ID', () => {
     // #given
-    const { onPlaylistSelect } = renderPlaylistSelection();
+    const { onPlaylistSelect } = renderLibraryPage();
 
     // #when
     fireEvent.click(screen.getByText('Chill Vibes'));
@@ -173,7 +185,7 @@ describe('PlaylistSelection', () => {
     });
 
     // #when
-    renderPlaylistSelection();
+    renderLibraryPage();
 
     // #then
     // When auth is resolved but initial load is not complete, the component renders
@@ -194,7 +206,7 @@ describe('PlaylistSelection', () => {
     });
 
     // #when
-    renderPlaylistSelection();
+    renderLibraryPage();
 
     // #then
     expect(
@@ -204,7 +216,7 @@ describe('PlaylistSelection', () => {
 
   it('Liked Songs item is present with LIKED_SONGS_ID', () => {
     // #given
-    const { onPlaylistSelect } = renderPlaylistSelection();
+    const { onPlaylistSelect } = renderLibraryPage();
 
     // #when
     fireEvent.click(screen.getByText('Liked Songs'));
@@ -214,9 +226,9 @@ describe('PlaylistSelection', () => {
     expect(onPlaylistSelect).toHaveBeenCalledWith(LIKED_SONGS_ID, 'Liked Songs', undefined);
   });
 
-  it('inDrawer: tapping a playlist opens the menu; Play then calls onPlaylistSelect', () => {
+  it('DrawerLibrary: tapping a playlist opens the menu; Play then calls onPlaylistSelect', () => {
     // #given
-    const { onPlaylistSelect } = renderPlaylistSelection({ inDrawer: true });
+    const { onPlaylistSelect } = renderDrawerLibrary();
 
     // #when
     fireEvent.click(screen.getByText('Chill Vibes'));
@@ -239,7 +251,7 @@ describe('PlaylistSelection — search and filter', () => {
 
   it('filters playlists by search query', () => {
     // #given
-    renderPlaylistSelection();
+    renderLibraryPage();
     const searchInput = screen.getByPlaceholderText('Search playlists...');
 
     // #when
@@ -253,7 +265,7 @@ describe('PlaylistSelection — search and filter', () => {
 
   it('shows empty state when search matches nothing', () => {
     // #given
-    renderPlaylistSelection();
+    renderLibraryPage();
     const searchInput = screen.getByPlaceholderText('Search playlists...');
 
     // #when
@@ -266,7 +278,7 @@ describe('PlaylistSelection — search and filter', () => {
 
   it('clears search when clear button is clicked', () => {
     // #given
-    renderPlaylistSelection();
+    renderLibraryPage();
     const searchInput = screen.getByPlaceholderText('Search playlists...') as HTMLInputElement;
     fireEvent.change(searchInput, { target: { value: MOCK_PLAYLIST_NAMES.CHILL } });
     expect(searchInput.value).toBe(MOCK_PLAYLIST_NAMES.CHILL);
@@ -288,7 +300,7 @@ describe('PlaylistSelection — search and filter', () => {
 
   it('search is case-insensitive', () => {
     // #given
-    renderPlaylistSelection();
+    renderLibraryPage();
     const searchInput = screen.getByPlaceholderText('Search playlists...');
 
     // #when
@@ -308,7 +320,7 @@ describe('PlaylistSelection — search and filter', () => {
 
   it('search updates live as user types', () => {
     // #given
-    renderPlaylistSelection();
+    renderLibraryPage();
     const searchInput = screen.getByPlaceholderText('Search playlists...');
 
     // #when
